fix(chz): clear -Wfloat-equal in runner; document corner_hz physics for downstream
Task 8a review fixes: replace 6 float == sites (base-case selection, log-interp guard, crossing-slope guard, test host-rate check) with tolerance/int compares — build is warning-clean. Add a downstream note (Tasks 9-12) that corner_hz is the measured physical -3 dB point (~0.44*fc for an authentic 4-pole Moog at res=0, NOT the nominal cutoff) and slope_db_oct is a transition-band reading. Reviewer confirmed corner=441 Hz is authentic (0.435*fc theory + 1.4% ZDF warp, 0.206 dB dual-method agreement), not an extraction bug.

+++ tests/CharacterizationRunnerTests.cpp
@@ -8,7 +8,7 @@ struct CharacterizationRunnerTests : public juce::UnitTest {
         beginTest("coarseGrid is bounded: 96k only, OS {1,2,4,8}, live only");
         {
             auto g = chz::coarseGrid();
-            expect(g.hostRates.size() == 1 && g.hostRates[0] == 96000.0, "single host rate 96k");
+            expect(g.hostRates.size() == 1 && (int) g.hostRates[0] == 96000, "single host rate 96k");
             expect(g.osFactors == (std::vector<int>{1,2,4,8}), "all four OS factors");
             expect(g.osModes.size() == 1 && g.osModes[0] == chz::OsMode::Live, "live only");
         }
